finalise multi images
