+++ htdocs/modules/system/templates/system_imagemanager2.tpl
@@ -141,7 +141,7 @@
         element: document.getElementById('fine-uploader-manual-trigger'),
         template: 'qq-template-manual-trigger',
         request: {
-            endpoint: '<{$xoops_url}>/Frameworks/fine-uploader/endpoint.php?imgcat_id=<{$show_cat}>'
+            endpoint: '<{$xoops_url}>/imagemanager.php?op=doupload&imgcat_id=<{$show_cat}>'
         },
         thumbnails: {
             placeholders: {
